// web-react/src/rpc/client.gen.ts
export interface Strum {
  label?: string
  strokes: string
  beats: string
}